requirements.txt
viam-sdk==0.11.0
google-api-python-client==2.0.0
pyinstaller==6.3.0